Changed secret store read return data type (#518)
- The sest_read function now returns a size_t value, for consistency with all other SGX HAL interfaces
- Updated dependent uses
- Updated unit tests

+++ firmware/src/hal/sgx/src/trusted/secret_store.h
@@ -56,7 +56,7 @@ bool sest_exists(char* key);
  *
  * @returns the length of the secret read, or ZERO upon error
  */
-uint8_t sest_read(char* key, uint8_t* dest, size_t dest_length);
+size_t sest_read(char* key, uint8_t* dest, size_t dest_length);
 
 /**
  * @brief Write a secret to the store
